requirements.txt
flask
gunicorn
python-dotenv
edge-tts
art
supabase
